# src/api/admin/role/dto/create/create.request.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEnum, IsNotEmpty, IsOptional, IsString } from 'class-validator';

import { PolicyRoleType } from '~/common/policy/constants';

export class CreateRoleDto {
  @ApiProperty({ required: true })
  @IsString()
  @IsNotEmpty()
  name: string;

  @ApiProperty({ required: false })
  @IsString()
  @IsOptional()
  description?: string;

  @ApiProperty({ required: true, enum: PolicyRoleType })
  @IsEnum(PolicyRoleType)
  type: PolicyRoleType;
}