# swagger.json
{"openapi":"3.0.0","paths":{"/api/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}},"/api/health/memory":{"get":{"operationId":"HealthController_checkMemory","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}},"/api/admin/auth/register":{"post":{"operationId":"AdminAuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterAdminPayloadDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginWithCredentialsResponseDto"}}}},"201":{"description":""}},"tags":["AdminAuth"]}},"/api/admin/auth/login/credentials":{"post":{"operationId":"AdminAuthController_loginWithCredentials","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginWithCredentialsPayloadDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginWithCredentialsResponseDto"}}}},"201":{"description":""}},"tags":["AdminAuth"]}},"/api/admin/roles":{"post":{"operationId":"RoleController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":""}},"tags":["Admin Role"]},"get":{"operationId":"RoleController_list","parameters":[{"name":"page","required":true,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"default":10,"type":"number"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"type","required":false,"in":"query","schema":{"type":"string"}},{"name":"isActive","required":false,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleListResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["Admin Role"]}}},"info":{"title":"Fishing marketplace API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"refreshToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginWithCredentialsResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"expiresIn":{"type":"number"}},"required":["accessToken","refreshToken","expiresIn"]},"RegisterAdminPayloadDto":{"type":"object","properties":{"email":{"type":"string","example":"test@example.com"},"name":{"type":"string","example":"John"},"password":{"type":"string","example":"Qwerty_123"}},"required":["email","name","password"]},"LoginWithCredentialsPayloadDto":{"type":"object","properties":{}},"CreateRoleDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"type":{"type":"string","enum":["super_admin","admin","vendor","customer"]}},"required":["name","type"]},"PermissionDto":{"type":"object","properties":{"subject":{"type":"string","enum":["user","role","auth"]},"actions":{"type":"array","items":{"type":"string","enum":["manage","read","create","update","delete"]}}},"required":["subject","actions"]},"RoleDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"object","nullable":true},"isActive":{"type":"boolean"},"type":{"type":"string","enum":["super_admin","admin","vendor","customer"]},"permissions":{"type":"array","items":{"$ref":"#/components/schemas/PermissionDto"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","description","isActive","type","permissions","createdAt","updatedAt"]},"PaginationMetaDto":{"type":"object","properties":{"itemCount":{"type":"number"},"totalItems":{"type":"number"},"itemsPerPage":{"type":"number"},"totalPages":{"type":"number"},"currentPage":{"type":"number"}},"required":["itemCount","totalItems","itemsPerPage","totalPages","currentPage"]},"RoleListResponseDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/RoleDto"}},"pagination":{"$ref":"#/components/schemas/PaginationMetaDto"}},"required":["data","pagination"]},"ErrorResponseDto":{"type":"object","properties":{"type":{"type":"string","enum":["invalid_argument","invalid_data","not_found","unauthorized","forbidden","conflict","too_many_requests","service_unavailable"]},"message":{"type":"string"},"code":{"type":"number","nullable":true},"errors":{"type":"array","items":{"type":"string"}}},"required":["type","message","code","errors"]}}}}